Upgrade to .NET Standard 2.0
