+++ FileDownloader.sln
@@ -1,28 +1,48 @@
 ﻿
 Microsoft Visual Studio Solution File, Format Version 12.00
-# Visual Studio 14
-VisualStudioVersion = 14.0.25123.0
+# Visual Studio 15
+VisualStudioVersion = 15.0.28307.168
 MinimumVisualStudioVersion = 10.0.40219.1
-Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "FileDownloader", "src\FileDownloader.csproj", "{DEB25DF4-20AF-4F85-8920-C3E14E147D9D}"
+Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "FileDownloader.Standard", "FileDownloader.Standard\FileDownloader.Standard.csproj", "{4CD514AD-F3E1-4C08-8D6E-680921A21902}"
 EndProject
-Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "FileDownloader.UnitTest", "test\FileDownloader.UnitTest.csproj", "{8DC80AFA-4F40-4961-8A3F-D92F27CB7448}"
+Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "FileDownloader.Standard.UnitTest", "FileDownloader.Standard.UnitTest\FileDownloader.Standard.UnitTest.csproj", "{CF506DB8-EEE1-4C7B-98EE-97B5B2F2E102}"
+EndProject
+Project("{2150E333-8FDC-42A3-9474-1A3956D46DE8}") = "Solution Items", "Solution Items", "{32FA2B28-C0FD-4FA7-BB17-4A3D095961CA}"
+	ProjectSection(SolutionItems) = preProject
+		.gitignore = .gitignore
+		LICENSE = LICENSE
+		README.md = README.md
+	EndProjectSection
 EndProject
 Global
 	GlobalSection(SolutionConfigurationPlatforms) = preSolution
+		Debug|Any CPU = Debug|Any CPU
 		Debug|x86 = Debug|x86
+		Release|Any CPU = Release|Any CPU
 		Release|x86 = Release|x86
 	EndGlobalSection
 	GlobalSection(ProjectConfigurationPlatforms) = postSolution
-		{DEB25DF4-20AF-4F85-8920-C3E14E147D9D}.Debug|x86.ActiveCfg = Debug|x86
-		{DEB25DF4-20AF-4F85-8920-C3E14E147D9D}.Debug|x86.Build.0 = Debug|x86
-		{DEB25DF4-20AF-4F85-8920-C3E14E147D9D}.Release|x86.ActiveCfg = Release|x86
-		{DEB25DF4-20AF-4F85-8920-C3E14E147D9D}.Release|x86.Build.0 = Release|x86
-		{8DC80AFA-4F40-4961-8A3F-D92F27CB7448}.Debug|x86.ActiveCfg = Debug|x86
-		{8DC80AFA-4F40-4961-8A3F-D92F27CB7448}.Debug|x86.Build.0 = Debug|x86
-		{8DC80AFA-4F40-4961-8A3F-D92F27CB7448}.Release|x86.ActiveCfg = Release|x86
-		{8DC80AFA-4F40-4961-8A3F-D92F27CB7448}.Release|x86.Build.0 = Release|x86
+		{4CD514AD-F3E1-4C08-8D6E-680921A21902}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{4CD514AD-F3E1-4C08-8D6E-680921A21902}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{4CD514AD-F3E1-4C08-8D6E-680921A21902}.Debug|x86.ActiveCfg = Debug|Any CPU
+		{4CD514AD-F3E1-4C08-8D6E-680921A21902}.Debug|x86.Build.0 = Debug|Any CPU
+		{4CD514AD-F3E1-4C08-8D6E-680921A21902}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{4CD514AD-F3E1-4C08-8D6E-680921A21902}.Release|Any CPU.Build.0 = Release|Any CPU
+		{4CD514AD-F3E1-4C08-8D6E-680921A21902}.Release|x86.ActiveCfg = Release|Any CPU
+		{4CD514AD-F3E1-4C08-8D6E-680921A21902}.Release|x86.Build.0 = Release|Any CPU
+		{CF506DB8-EEE1-4C7B-98EE-97B5B2F2E102}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{CF506DB8-EEE1-4C7B-98EE-97B5B2F2E102}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{CF506DB8-EEE1-4C7B-98EE-97B5B2F2E102}.Debug|x86.ActiveCfg = Debug|Any CPU
+		{CF506DB8-EEE1-4C7B-98EE-97B5B2F2E102}.Debug|x86.Build.0 = Debug|Any CPU
+		{CF506DB8-EEE1-4C7B-98EE-97B5B2F2E102}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{CF506DB8-EEE1-4C7B-98EE-97B5B2F2E102}.Release|Any CPU.Build.0 = Release|Any CPU
+		{CF506DB8-EEE1-4C7B-98EE-97B5B2F2E102}.Release|x86.ActiveCfg = Release|Any CPU
+		{CF506DB8-EEE1-4C7B-98EE-97B5B2F2E102}.Release|x86.Build.0 = Release|Any CPU
 	EndGlobalSection
 	GlobalSection(SolutionProperties) = preSolution
 		HideSolutionNode = FALSE
+	EndGlobalSection
+	GlobalSection(ExtensibilityGlobals) = postSolution
+		SolutionGuid = {225D593E-C21C-406D-8941-9FFB8BB8E052}
 	EndGlobalSection
 EndGlobal
